requirements.txt
beautifulsoup4==4.7.1
certifi==2017.4.17
chardet==3.0.4
future==0.16.0
google-api-python-client==1.5.5
HTMLParser==0.0.2
httplib2==0.9.2
idna==2.5
oauth2client==4.0.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
PyDrive==1.3.1
python-telegram-bot==11.1.0
PyYAML==3.12
requests==2.18.1
rsa==3.4.2
six==1.10.0
soupsieve==1.9.1
uritemplate==3.0.0
urllib3==1.21.1
